remove displayHeader and rename duel-columns
