--- src/components/content-display/new-article/collection/stack/styles/display-header.module.scss
@@ -1,6 +1,0 @@
-.displayHeader {
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	gap: 1rem;
-}
